chatbot-gpt2-service/requirements.txt
fastapi
uvicorn
torch
transformers
python-dotenv
